partitioned writer group of functions
